+++ src/main/kotlin/kr/hhplus/be/server/domain/coupon/CouponEventListener.kt
@@ -6,7 +6,7 @@ import org.springframework.kafka.support.Acknowledgment
 import org.springframework.stereotype.Component
 
 @Component
-class CouponSelfEventListener(
+class CouponEventListener(
     private val couponService: CouponService
 ) {
