+++ src/main/kotlin/kr/hhplus/be/server/interfaces/event/coupon/CouponExternalEventListener.kt
@@ -6,14 +6,14 @@ import org.springframework.context.event.EventListener
 import org.springframework.stereotype.Component
 
 @Component
-class CouponEventListener(
+class CouponExternalEventListener(
     private val couponService: CouponService
 ) {
 
     @EventListener
     fun handle(event: ProductEvent.Completed) {
         println("🔥 이벤트 수신됨: $event")
-        val command = CouponEventMapper.toCommand(event)
+        val command = CouponExternalEventMapper.toCommand(event)
         couponService.isUsed(command)
     }
