+++ src/main/kotlin/kr/hhplus/be/server/interfaces/event/coupon/CouponExternalEventMapper.kt
@@ -1,12 +1,12 @@
-package kr.hhplus.be.server.interfaces.event.point
+package kr.hhplus.be.server.interfaces.event.coupon
 
-import kr.hhplus.be.server.domain.coupon.CouponEvent
-import kr.hhplus.be.server.domain.point.PointCommand
+import kr.hhplus.be.server.domain.coupon.CouponCommand
+import kr.hhplus.be.server.domain.product.ProductEvent
 
-class PointEventMapper {
+class CouponExternalEventMapper {
     companion object {
-        fun toCommand(event: CouponEvent.Completed): PointCommand.Deduct {
-            return PointCommand.Deduct(
+        fun toCommand(event: ProductEvent.Completed): CouponCommand.Use {
+            return CouponCommand.Use(
                 orderId = event.orderId,
                 userId = event.userId,
                 pointId = event.pointId,
@@ -14,15 +14,13 @@ class PointEventMapper {
                 couponId = event.couponId,
                 totalPrice = event.totalPrice,
                 productsDetail = event.productsDetail.map {
-                    PointCommand.OrderedProduct(
+                    CouponCommand.OrderedProduct(
                         productId = it.productId,
                         name = it.name,
                         price = it.price,
                         quantity = it.quantity
                     )
-                },
-                discountType = event.discountType,
-                discountValue = event.discountValue,
+                }
             )
         }
     }
