+++ src/main/kotlin/kr/hhplus/be/server/interfaces/event/product/ProductExternalEventMapper.kt
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.interfaces.event.product
 import kr.hhplus.be.server.domain.order.OrderEvent
 import kr.hhplus.be.server.domain.product.ProductCommand
 
-class ProductEventMapper {
+class ProductExternalEventMapper {
     companion object {
         fun toCommand(event: OrderEvent.Create): ProductCommand.Deduct {
             return ProductCommand.Deduct(
