+++ src/main/kotlin/kr/hhplus/be/server/interfaces/event/rank/RankExternalEventMapper.kt
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.interfaces.event.rank
 import kr.hhplus.be.server.domain.order.OrderEvent
 import kr.hhplus.be.server.domain.rank.ProductRankingCommand
 
-class RankEventMapper {
+class RankExternalEventMapper {
     companion object {
         fun toCommand(event: OrderEvent.Completed): ProductRankingCommand.Increase {
             return ProductRankingCommand.of(
